+++ Lab_data.csv
@@ -1,0 +1,7 @@
+0.264286300000002,0.2377346082638394,5.651774396636118,1.1303548793272238
+0.5184105999999993,0.12119736749210008,1.468880188701516,0.3304980424578411
+0.5143966999999989,0.12214308528806686,1.4918933283687974,0.33567599888297944
+0.26500219999999786,0.23709237130861752,5.621279253274336,1.2647878319867256
+0.765646199999999,0.0820614012059357,0.6734073567881546,0.15151665527733477
+0.2673723000000052,0.2349906852729276,5.522062216504011,1.2424639987134025
+1.2668515999999954,0.049595390651912374,0.2459702773915797,0.05534331241310543
